+++ languages/cpp/parser/commentformatter.h
@@ -35,10 +35,10 @@ class ParseSession;
 class KDEVCPPPARSER_EXPORT CommentFormatter {
     public:
     ///Processes the comment represented by the given token-number within the parse-session's token-stream
-    static QByteArray formatComment( size_t token, const ParseSession* session );
+    static QByteArray formatComment( uint token, const ParseSession* session );
 
     ///Processes the list of comments represented by the given token-number within the parse-session's token-stream
-    static QByteArray formatComment( const ListNode<size_t>* node, const ParseSession* session );
+    static QByteArray formatComment( const ListNode<uint>* node, const ParseSession* session );
   private:
 };
